Moved AP-AssetBuilder specific types into AssetBuilder.Static library.  Also removed some unused/old code
