+++ Code/Tools/AssetProcessor/AssetBuilder/AssetBuilderComponent.cpp
@@ -34,6 +34,7 @@
 #include <AzCore/Interface/Interface.h>
 #include <AzFramework/Asset/AssetSystemComponent.h>
 #include <ToolsComponents/ToolsAssetCatalogComponent.h>
+#include <AssetBuilderStatic.h>
 
 // Command-line parameter options:
 static const char* const s_paramHelp = "help"; // Print help information.
@@ -55,7 +56,6 @@ static const char* const s_paramRegisterBuilders = "register"; // Indicates the 
 
 // Task modes:
 static const char* const s_taskResident = "resident"; // stays up and running indefinitely, accepting jobs via network connection
-static const char* const s_taskRegisterBuilder = "register"; // outputs all the builder descriptors
 static const char* const s_taskCreateJob = "create"; // runs a builders createJobs function
 static const char* const s_taskProcessJob = "process"; // runs processJob function
 static const char* const s_taskDebug = "debug"; // runs a one shot job in a fake environment for a specified file.
@@ -207,7 +207,7 @@ void AssetBuilderComponent::Reflect(AZ::ReflectContext* context)
 
 bool AssetBuilderComponent::DoHelloPing()
 {
-    using namespace AssetBuilderSDK;
+    using namespace AssetBuilder;
 
     BuilderHelloRequest request;
     BuilderHelloResponse response;
@@ -409,11 +409,11 @@ bool AssetBuilderComponent::ConnectToAssetProcessor()
 
 bool AssetBuilderComponent::SendRegisteredBuildersToAp()
 {
-    AssetBuilderSDK::BuilderRegistrationRequest registrationRequest;
+    AssetBuilder::BuilderRegistrationRequest registrationRequest;
 
     for (const auto& [uuid, desc] : m_assetBuilderDescMap)
     {
-        AssetBuilderSDK::BuilderRegistration registration;
+        AssetBuilder::BuilderRegistration registration;
 
         registration.m_name = desc->m_name;
         registration.m_analysisFingerprint = desc->m_analysisFingerprint;
@@ -436,7 +436,7 @@ bool AssetBuilderComponent::SendRegisteredBuildersToAp()
 
 bool AssetBuilderComponent::RunInResidentMode(bool sendRegistration)
 {
-    using namespace AssetBuilderSDK;
+    using namespace AssetBuilder;
     using namespace AZStd::placeholders;
 
     AZ_TracePrintf("AssetBuilderComponent", "RunInResidentMode: Starting resident mode (waiting for commands to arrive)\n");
@@ -775,11 +775,7 @@ bool AssetBuilderComponent::RunOneShotTask(const AZStd::string& task)
 
     AZ::StringFunc::Path::Normalize(inputFilePath);
     AZ::StringFunc::Path::Normalize(outputFilePath);
-    if (task == s_taskRegisterBuilder)
-    {
-        return HandleRegisterBuilder(inputFilePath, outputFilePath);
-    }
-    else if (task == s_taskCreateJob)
+    if (task == s_taskCreateJob)
     {
         auto func = [this](const AssetBuilderSDK::CreateJobsRequest& request, AssetBuilderSDK::CreateJobsResponse& response)
             {
@@ -935,7 +931,7 @@ void AssetBuilderComponent::JobThread()
         {
             case JobType::Create:
             {
-                using namespace AssetBuilderSDK;
+                using namespace AssetBuilder;
 
                 auto* netRequest = azrtti_cast<CreateJobsNetRequest*>(job->m_netRequest.get());
                 auto* netResponse = azrtti_cast<CreateJobsNetResponse*>(job->m_netResponse.get());
@@ -961,7 +957,7 @@ void AssetBuilderComponent::JobThread()
             }
             case JobType::Process:
             {
-                using namespace AssetBuilderSDK;
+                using namespace AssetBuilder;
 
                 AZ_TracePrintf("AssetBuilder", "Running processJob task\n");
 
@@ -1020,14 +1016,14 @@ void AssetBuilderComponent::JobThread()
 
 void AssetBuilderComponent::CreateJobsResidentHandler(AZ::u32 /*typeId*/, AZ::u32 serial, const void* data, AZ::u32 dataLength)
 {
-    using namespace AssetBuilderSDK;
+    using namespace AssetBuilder;
 
     ResidentJobHandler<CreateJobsNetRequest, CreateJobsNetResponse>(serial, data, dataLength, JobType::Create);
 }
 
 void AssetBuilderComponent::ProcessJobResidentHandler(AZ::u32 /*typeId*/, AZ::u32 serial, const void* data, AZ::u32 dataLength)
 {
-    using namespace AssetBuilderSDK;
+    using namespace AssetBuilder;
 
     ResidentJobHandler<ProcessJobNetRequest, ProcessJobNetResponse>(serial, data, dataLength, JobType::Process);
 }
@@ -1055,18 +1051,6 @@ bool AssetBuilderComponent::HandleTask(const AZStd::string& inputFilePath, const
     }
 
     return true;
-}
-
-bool AssetBuilderComponent::HandleRegisterBuilder(const AZStd::string& /*inputFilePath*/, const AZStd::string& outputFilePath) const
-{
-    AssetBuilderSDK::RegisterBuilderResponse response;
-
-    for (const auto& pair : m_assetBuilderDescMap)
-    {
-        response.m_assetBuilderDescList.push_back(*pair.second);
-    }
-
-    return AZ::Utils::SaveObjectToFile(outputFilePath, AZ::DataStream::ST_XML, &response);
 }
 
 void AssetBuilderComponent::UpdateResultCode(const AssetBuilderSDK::ProcessJobRequest& request, AssetBuilderSDK::ProcessJobResponse& response) const
