+++ Code/Tools/AssetProcessor/AssetBuilder/asset_builder_static_files.cmake
@@ -1,0 +1,12 @@
+#
+# Copyright (c) Contributors to the Open 3D Engine Project.
+# For complete copyright and license terms please see the LICENSE at the root of this distribution.
+#
+# SPDX-License-Identifier: Apache-2.0 OR MIT
+#
+#
+
+set(FILES
+    AssetBuilderStatic.h
+    AssetBuilderStatic.cpp
+)
